# Avoid rehashing keys in transform_values