Previously, calling transform_values would call `rb_hash_aset` for each key, needing to rehash it and look up its location.

Instead, we can use `rb_hash_stlike_foreach_with_replace` to replace the values as we iterate without rehashing the keys.

**Before**

```
$ time ./ruby --disable-gems -e 'h = Hash[(?a..?z).map{|x| [x,1] }]; 1_000_000.times{ h.transform_values!{|x| x } }'
./ruby --disable-gems -e   2.04s user 0.02s system 97% cpu 2.178 total
$ irb
irb(main):001:0> o = Object.new
irb(main):002:0> def o.hash; puts "hash"; 5; end
=> :hash
irb(main):003:0> h = {o => 1}
hash
irb(main):004:0> h.transform_values{|x| x + 1 }
hash
=> {#<Object:0x00007fa19ca74208>=>2}
```

**After**

```
$ time ./ruby --disable-gems -e 'h = Hash[(?a..?z).map{|x| [x,1] }]; 1_000_000.times{ h.transform_values!{|x| x } }'
./ruby --disable-gems -e   1.07s user 0.02s system 98% cpu 1.104 total
$ irb
irb(main):001:0> o = Object.new
irb(main):002:0> def o.hash; puts "hash"; 5; end
=> :hash
irb(main):003:0> h = {o => 1}
hash
irb(main):004:0> h.transform_values{|x| x + 1 }
=> {#<Object:0x00007fdcdb917bd8>=>2}
```

([additional benchmarks](https://gist.github.com/jhawthorn/591e27f4e1af8cf9694d6629c8e6dd9e))